+++ lib/auth.ts
@@ -1,0 +1,8 @@
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { getServerSession } from "next-auth";
+
+
+export const auth = async () => {
+  const session = await getServerSession(authOptions);
+  return session;
+};
